backend/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
python-jose[cryptography]
python-multipart
python-dotenv
passlib
bcrypt>=4.0.0
email-validator
